+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Nanook::Account#info` now also returns the `id` of the account
 - `Nanook::Block#info` now also returns the `id` of the block
 
+### Fixed
+- Missing `Nanook#rpc` accessor
+
 ## 1.0.1
 
 ### Fixed